requirements.txt
pandas~=2.2.2
numpy~=1.26.4
Flask~=3.0.0
python-dateutil~=2.9.0.post0
psutil~=5.9.8
requests~=2.31.0
pytz~=2024.1
selenium~=4.20.0
openai~=1.13.3
streamlit~=1.35.0
faiss-cpu==1.8.0
GitPython==3.1.43
requests==2.31.0
python-dotenv~=1.0.0